rename checker, add some checks & exceptions, simplify code
