+++ cmd/gocritic/testdata/boolFuncPrefix/negative_tests.go
