+++ cmd/gocritic/testdata/boolFuncPrefix/positive_tests.go
