+++ docs/checkers/boolFuncPrefix.md
